refactor(channels): split the template create form out of the manager
The manager was 570 lines doing two jobs, and authentication templates
need a category-dependent form that would push it further. Move the
form to its own component first, with no behaviour change, so that
change lands somewhere reviewable.

Manager keeps the shell, the list and delete (452 lines); the form owns
its own state and submit (206).
